Changelog and currency validation function
